현재유저 필요한부분수정완

+++ src/main/java/com/jobdam/sns/dto/SnsMessageBoxResponseDto.java
@@ -7,14 +7,14 @@ import java.time.LocalDateTime;
 @Getter
 public class SnsMessageBoxResponseDto {
 
-    private Long opponentUserId;
+    private Integer opponentUserId;
     private String opponentNickname;
     private String opponentProfileImageUrl;
     private String lastMessageContent;
     private LocalDateTime lastMessageCreatedAt;
 
     public SnsMessageBoxResponseDto(
-            Long opponentUserId,
+            Integer opponentUserId,
             String opponentNickname,
             String opponentProfileImageUrl,
             String lastMessageContent,
